requirements.txt
asgiref==3.8.1
Django==5.0.3
pip==23.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
djangorestframework==3.15.1
